renames samfundet logo used in Logo.tsx component and moved Org context into context folder. Other refactoring to import these changes correctly
