+++ frontend/src/Components/Logo/components/index.ts
@@ -1,3 +1,3 @@
 export { IsfitLogo } from './IsfitLogo';
-export { SamfLogo } from './SamfLogo';
-export { UkaLogo } from './UkaLogo';
+export { SamfundetLogo } from './SamfundetLogo';
+export { UKALogo } from './UKALogo';
